requirements.txt
# Created from pipreqs (command: "pipreqs .")
click
torch>=1.7.0
torchvision
humanize
matplotlib
colorlog
prettytable
requests
tqdm
numpy
Pillow
ipython
scipy
httplib2
oauth2client
google-api-python-client
pydrive
tensorboard
pyyaml
h5py
yaml